backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
tzdata==2025.2
openpyxl==3.1.2
python-dateutil==2.8.2
